Fit a pretrained lasso model using glmnet.

Fits a pretrained lasso model using the glmnet package, for a fixed choice of the pretraining hyperparameter alpha. Additionally fits an "overall" model (using all data) and "individual" models (use each individual group). Can fit input-grouped data with Gaussian, multinomial, binomial or Cox outcomes, and target-grouped data, which necessarily has a multinomial outcome. Many ptLasso arguments are passed directly to glmnet, and therefore the glmnet documentation is another good reference for ptLasso.

Usage

ptLassoMult(
  x,
  y,
  alpha = 0.5,
  type.measure = c("default", "mse", "mae", "deviance"),
  overall.lambda = c("lambda.1se", "lambda.min"),
  overall.gamma = "gamma.1se",
  foldid = NULL,
  nfolds = 10,
  standardize = TRUE,
  verbose = FALSE,
  weights = NULL,
  penalty.factor = rep(1, nvars),
  fitoverall = NULL,
  fitind = NULL,
  en.alpha = 1,
  ...
)

Arguments

x

input matrix, of dimension nobs x nvars; each row is an observation vector. Can be in sparse matrix format (inherit from class '"sparseMatrix"' as in package 'Matrix'). Requirement: 'nvars >1'; in other words, 'x' should have 2 or more columns.

y

quantitative response variable, of dimension nobs x nresponses.

alpha

The pretrained lasso hyperparameter, with \(0\le\alpha\le 1\). The range of alpha is from 0 (which fits the overall model with fine tuning) to 1 (the individual models). The default value is 0.5, chosen mostly at random. To choose the appropriate value for your data, please either run ptLasso with a few choices of alpha and evaluate with a validation set, or use cv.ptLasso, which recommends a value of alpha using cross validation.

type.measure

loss to use for cross-validation within each individual, overall, or pretrained lasso model. Choices are 'type.measure="mse"' (mean squared error), 'type.measure="mae"' (mean absolute error) and 'type.measure="deviance"'.

overall.lambda

The choice of lambda to be used by the overall model to define the offset and penalty factor for pretrained lasso. Defaults to "lambda.1se", could alternatively be "lambda.min". This choice of lambda will be used to compute the offset and penalty factor (1) during model training and (2) during prediction. In the predict function, another lambda must be specified for the individual models, the second stage of pretraining and the overall model.

overall.gamma

For use only when the option relax = TRUE is specified. The choice of gamma to be used by the overall model to define the offset and penalty factor for pretrained lasso. Defaults to "gamma.1se", but "gamma.min" is also a good option. This choice of gamma will be used to compute the offset and penalty factor (1) during model training and (2) during prediction. In the predict function, another gamma must be specified for the individual models, the second stage of pretraining and the overall model.

foldid

An optional vector of values between 1 and nfolds identifying what fold each observation is in. If supplied, nfold can be missing.

nfolds

Number of folds for CV (default is 10). Although nfoldscan be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. Smallest value allowable is nfolds = 3.

standardize

Should the predictors be standardized before fitting (default is TRUE).

verbose

If verbose=1, print a statement showing which model is currently being fit with cv.glmnet.

weights

observation weights. Default is 1 for each observation.

penalty.factor

Separate penalty factors can be applied to each coefficient. This is a number that multiplies 'lambda' to allow differential shrinkage. Can be 0 for some variables, which implies no shrinkage, and that variable is always included in the model. Default is 1 for all variables (and implicitly infinity for variables listed in 'exclude'). For more information, see ?glmnet. For pretraining, the user-supplied penalty.factor will be multiplied by the penalty.factor computed by the overall model.

fitoverall

An optional cv.glmnet object specifying the overall model. This should have been trained on the full training data, with the argumnet keep = TRUE.

fitind

An optional list of cv.glmnet objects specifying the individual models.

en.alpha

The elasticnet mixing parameter, with 0 <= en.alpha <= 1. The penalty is defined as (1-alpha)/2||beta||_2^2+alpha||beta||_1. 'alpha=1' is the lasso penalty, and 'alpha=0' the ridge penalty. Default is `en.alpha = 1` (lasso).

...

Additional arguments to be passed to the cv.glmnet functions. Notable choices include "trace.it" and "parallel". If trace.it = TRUE, then a progress bar is displayed for each call to "cv.glmnet"; useful for big models that take a long time to fit. If parallel = TRUE, use parallel foreach to fit each fold. Must register parallel before hand, such as doMC or others. ptLasso does not support the arguments intercept, offset, fit and check.args.

Value

An object of class "ptLasso", which is a list with the ingredients of the fitted models.

call

The call that produced this object.

alpha

The value of alpha used for pretraining.

fitoverall

A fitted cv.glmnet object trained using the full data.

fitpre

A list of fitted (pretrained) cv.glmnet objects, one trained for each response.

fitind

A list of fitted cv.glmnet objects, one trained for each response.

fitoverall.lambda

Lambda used with fitoverall, to compute the offset for pretraining.

References

Friedman, J., Hastie, T., & Tibshirani, R. (2010). Regularization paths for generalized linear models via coordinate descent. Journal of Statistical Software, 33(1), 1-22.

See also

Examples

# Getting started. First, we simulate data: we need covariates x and multiresponse y.
set.seed(1234)
n = 1000; ntrain = 500;
p = 500
sigma = 2
     
x = matrix(rnorm(n*p), n, p)
beta1 = c(rep(1, 5), rep(0.5, 5), rep(0, p - 10))
beta2 = c(rep(1, 5), rep(0, 5), rep(0.5, 5), rep(0, p - 15))

mu = cbind(x %*% beta1, x %*% beta2)
y  = cbind(mu[, 1] + sigma * rnorm(n), 
           mu[, 2] + sigma * rnorm(n))
cat("SNR for the two tasks:", round(diag(var(mu)/var(y-mu)), 2), fill=TRUE)
#> SNR for the two tasks: 1.6 1.44
cat("Correlation between two tasks:", cor(y[, 1], y[, 2]), fill=TRUE)
#> Correlation between two tasks: 0.5164748

xtest = x[-(1:ntrain), ]
ytest = y[-(1:ntrain), ]

x = x[1:ntrain, ]
y = y[1:ntrain, ]

# Now, we can fit a ptLasso multiresponse model:
fit = ptLassoMult(x, y, alpha = 0.5, type.measure = "mse")
# plot(fit) # to see all of the cv.glmnet models trained
predict(fit, xtest) # to predict on new data
#> 
#> Call:  
#> predict.ptLasso(object = fit, xtest = xtest) 
#> 
#> 
#> alpha =  0.5 
#> 
#> Support size:
#>                                          
#> Overall    57                            
#> Pretrain   29 (19 common + 10 individual)
#> Individual 80                            
predict(fit, xtest, ytest=ytest) # if ytest is included, we also measure performance
#> 
#> Call:  
#> predict.ptLasso(object = fit, xtest = xtest, ytest = ytest) 
#> 
#> 
#> 
#> alpha =  0.5 
#> 
#> Performance (Mean squared error):
#> 
#>            allGroups  mean response_1 response_2
#> Overall        9.394 4.697      4.227      5.168
#> Pretrain       9.022 4.511      4.144      4.878
#> Individual     9.465 4.733      4.243      5.222
#> 
#> Support size:
#>                                          
#> Overall    57                            
#> Pretrain   29 (19 common + 10 individual)
#> Individual 80